+++ CHANGES.md
@@ -6,6 +6,8 @@ Markdown Writer FX Change Log
 - Project specific options: Select "Store in project" checkbox in Options dialog
   to save options in per project (in file `<project-root>/.markdownwriterfx`).
 - "Format only modified paragraphs on Save" checkbox added to Options dialog.
+- Fixed behavior of `Ctrl+RIGHT`, which now moves caret to beginning of next
+  word.
 - Disabled "Save As" for read-only editors (binary or too-large files).